api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
anthropic>=0.40.0
python-dotenv==1.0.1
